# defektor/server/src/main/resources/defektorOpenAPISpec/defektorOpenAPISpec.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "https://virtserver.swaggerhub.com/jaimelive/defektorOpenAPISpec/1.0.0"
    }
  ],
  "info": {
    "description": "defektor controll aplication interface",
    "version": "1.0.0",
    "title": "defektor OpenAPI Specification",
    "contact": {
      "email": "sob@dei.uc.pt"
    }
  },
  "paths": {
    "/slave": {
      "get": {
        "summary": "list slave machines",
        "operationId": "slaveList",
        "description": "List available slave machines",
        "responses": {
          "200": {
            "description": "slaves listed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Slave"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "add slave machine",
        "operationId": "slaveAdd",
        "description": "Add slave machine",
        "responses": {
          "201": {
            "description": "slave created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slave"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "slave already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Slave"
              }
            }
          },
          "description": "Added Slave"
        }
      }
    },
    "/slave/{slaveId}": {
      "get": {
        "summary": "slave machine info",
        "operationId": "slaveGet",
        "description": "Get slave machine information",
        "parameters": [
          {
            "in": "path",
            "name": "slaveId",
            "description": "Slave machine identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID",
              "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "slave information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slave"
                }
              }
            }
          },
          "400": {
            "description": "slave does not exist"
          }
        }
      },
      "delete": {
        "summary": "delete slave machine",
        "operationId": "slaveDelete",
        "description": "Delete slave machine",
        "parameters": [
          {
            "in": "path",
            "name": "slaveId",
            "description": "Slave machine identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "UUID",
              "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "slave deleted"
          },
          "400": {
            "description": "slave does not exist"
          }
        }
      }
    },
...
    "/target/{target}": {
      "get": {
        "summary": "list target instances",
        "operationId": "targetGet",
        "description": "List available targets",
        "parameters": [
          {
            "in": "path",
            "name": "target",
            "description": "target type",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "slaves listed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Target"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/system/config": {
      "get": {
        "summary": "list system configs",
        "operationId": "systemConfigList",
        "description": "List available system configs",
        "responses": {
          "200": {
            "description": "list of system configs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit system and its configurations",
        "operationId": "systemTypeConfigure",
        "description": "Submits a system type and its configuration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfig"
              }
            }
          },
          "description": "Configure system"
        },
        "responses": {
          "200": {
            "description": "submits a system type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfig"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TargetType": {
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "container"
          }
        }
      },
      "Target": {
        "required": [
          "type",
          "name"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/TargetType"
          },
          "name": {
            "type": "string",
            "example": "istio_ingress_2314234h21345"
          }
        }
      },
      "SSHCredentials": {
        "required": [
          "username",
          "key"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "debian"
          },
          "key": {
            "type": "string",
            "example": "ooooohhhh this is a super secret private key... 🔑"
          }
        }
      },
      "Slave": {
        "required": [
          "id",
          "address",
          "port",
          "credentials"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "UUID",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "address": {
            "type": "string",
            "example": "example.org"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "default": 22,
            "example": 22
          },
          "credentials": {
            "$ref": "#/components/schemas/SSHCredentials"
          }
        }
      },
...
      "SystemType": {
        "properties": {
          "name": {
            "type": "string",
            "example": "kubernetes"
          }
        }
      },
      "SystemConfig": {
        "properties": {
          "configs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyValue"
            }
          },
          "systemType": {
            "$ref": "#/components/schemas/SystemType"
          }
        }
      },
...
      "KeyValue": {
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      }
    }
  }
}